chore(app): refactor app struct
