+++ public/index.html
@@ -21,7 +21,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Poppy App</title>
+    <title></title>
 
     <script data-app-config type="application/json">
     {
