+++ src/index.tsx
@@ -1,17 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from 'src/widgets/App/App';
+import { App } from 'src/widgets/App/App';
 
-import { initApp } from 'src/utils/initApp';
-
-initApp()
-  .then(() => {
-    ReactDOM.render(<App />, document.getElementById('root'));
-  })
-  .catch((err) => {
-    console.error('app bootstrap error' + err?.stack);
-  });
+ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
