+++ src/widgets/App/App.module.css
@@ -7,3 +7,7 @@
 :global(body) {
   margin: 0;
 }
+
+.host {
+  height: 100%;
+}
