+++ src/widgets/App/App.tsx
@@ -5,10 +5,16 @@ import { Layout } from 'src/pages/Home/Home';
 import { useAppReadyInfo } from '../../utils/appReadyInfo';
 import { AppEmptyState } from '../../components/AppEmptyState/AppEmptyState';
 
-import './App.module.css';
+import styles from './App.module.css';
 import { AppLoading } from '../../components/AppLoading/AppLoading';
 
-export default function App(): JSX.Element | null {
+interface AppProps {
+  isWrapper?: boolean;
+  children?: React.ReactNode;
+}
+
+export const App: React.FC<AppProps> = (props: AppProps) => {
+  const isWrapper = props.isWrapper ?? false;
   const appReadyInfoState = useAppReadyInfo();
 
   if (appReadyInfoState.loading) {
@@ -17,11 +23,15 @@ export default function App(): JSX.Element | null {
 
   const appInfo = appReadyInfoState.appReadyInfo?.appInfo;
   if (!appInfo || appInfo.status !== 'normal' || appInfo.isExpired) {
-    return <AppEmptyState expired={!!appInfo?.isExpired} error={!!appReadyInfoState.error} disabled={appInfo?.status !== 'normal'} />;
+    return <AppEmptyState expired={!!appInfo?.isExpired} error={!!appReadyInfoState.error} disabled={!!appInfo?.status && appInfo?.status !== 'normal'} />;
+  }
+
+  if (isWrapper) {
+    return <div className={styles.host}>{props.children}</div>;
   }
 
   return (
-    <>
+    <div className={styles.host}>
       <HashRouter>
         <Switch>
           <Route path="/login" component={Login} />
@@ -29,6 +39,6 @@ export default function App(): JSX.Element | null {
           {/*<Redirect to="/" exact/>*/}
         </Switch>
       </HashRouter>
-    </>
+    </div>
   );
-}
+};
