--- src/widgets/StoryBookApp/StoryBookApp.module.css
@@ -1,4 +1,0 @@
-@import '~antd/dist/antd.css';
-
-.host {
-}
