--- src/widgets/StoryBookApp/StoryBookApp.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import styles from './StoryBookApp.module.css';
-
-export const StoryBookApp: React.FC = (props) => {
-  return <div>{props.children}</div>;
-};
